fixes in weapon and operator creation and increment in weapon options

+++ rainbow-six-siege/src/main/java/virionis/nikolas/rainbow/six/siege/attachments/Sight.java
@@ -1,13 +1,13 @@
 package virionis.nikolas.rainbow.six.siege.attachments;
 
 public class Sight extends Attachment {
-    private Integer zoom;
+    private Double zoom;
 
-    public Integer getZoom() {
+    public Double getZoom() {
         return zoom;
     }
 
-    public Sight(Integer zoom, String name) {
+    public Sight(Double zoom, String name) {
         super(name);
         this.zoom = zoom;
     }
